Move event detail field components into separate folder
