+++ src/components/EventDetails/DetailFields/RadiationDetailFields.tsx
@@ -1,15 +1,15 @@
 import React, { useMemo } from "react";
 import { FormikValues, useFormikContext } from "formik";
 
-import TextField from "../form/TextField";
-import { useGlobalContext } from "../../contexts/GlobalContext"
-import { GlobalValues } from "../../types/Global"
+import TextField from "../../form/TextField";
+import { useGlobalContext } from "../../../contexts/GlobalContext"
+import { GlobalValues } from "../../../types/Global"
 
-import { RadiationDetailFields as RadiationDetailFieldConfig } from "../../types/PatientEvent/Details/EventTypes/RadiationDetails";
-import DetailFieldsFactory from "../../types/PatientEvent/Details/DetailFieldsFactory";
+import { RadiationDetailFields as RadiationDetailFieldConfig } from "../../../types/PatientEvent/Details/EventTypes/RadiationDetails";
+import DetailFieldsFactory from "../../../types/PatientEvent/Details/DetailFieldsFactory";
 
 import prependDetailFieldName from "./prependDetailFieldName";
-import { Select } from "../form/FilterSelect";
+import { Select } from "../../form/FilterSelect";
 
 export default function RadiationDetailFields() {
   const { config }: GlobalValues = useGlobalContext();
