+++ src/components/EventDetails/DetailFields/prependDetailFieldName.ts
